requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.1
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
scikit-learn==1.5.1
xgboost==2.1.0
altair==5.3.0
google-genai>=0.2.0
ml-dtypes==0.3.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2